Plot Grid

This function plots the distribution of error scores resulting from gridSearch.

Usage

gridPlot(
  data,
  method = c("quantile", "normalized"),
  qcut1 = 0.05,
  qcut2 = 0.1,
  qcut3 = 0.2,
  ncut1 = 0.05,
  ncut2 = 0.1,
  ncut3 = 0.2
)

Arguments

data

A data frame generated by gridSearch

method

If method = quantile, the cuts are based on the quantile of the error distribution. If method = normalized, cuts are based on the normalized error scores (error / n), which ranges from 0 to 2, for pattern = contingency, or normalized KS statistics, which ranges from 0 to 1, for pattern = slopes.

qcut1

Cut 1 for the quantile method.

qcut2

Cut 2 for the quantile method.

qcut3

Cut 3 for the quantile method.

ncut1

Cut 1 for the normalized method.

ncut2

Cut 2 for the normalized method.

ncut3

Cut 3 for the normalized method.

Value

A ggplot2 object.