Grid Search for Adjudicating DGPs

This function uses a three-step grid search algorithm to calculate the extent to which a dataframe can be approximated by a known DGP. It implements fake-data simulations from a grid of plausible values—rate of change ranging from 0% to 100%, strength of change ranging from 0 SD to 2 SD, and 5 directional cases (everyone changing negatively, everyone changing positively, half changing negatively and half changing positively, 25% changing negatively and 75% changing positively, and 75% changing negatively and 25% changing positively), calculates the distribution of response patterns or slopes, and provides an error term that summarizes the distance of the DGP from the observed values.

Usage

gridSearch(
  data,
  yname,
  tname,
  pname,
  pattern = c("contingency", "slopes"),
  steps = 1,
  reliability = 0.9,
  seed = 11235,
  workers = 8
)

Arguments

data

A dataframe.

yname

The outcome identifier.

tname

The time identifier.

pname

The unit identifier.

pattern

If pattern = contingency, the calculations are based on the use of a contingency table of panel patterns; if pattern = slopes, function uses individual slope coefficients.

steps

The number of simulated datasets.

reliability

The assumed reliability score of the outcome measurement.

seed

Seed for reproducibility.

workers

The number of workers for parallelization (note that parallelization is highly recommended for reasonable duration).

Value

A data frame.

Examples

Basic gridSearch() call:

set.seed(1871)
data <- buildDGP(n = 50,
                 t = 3,
                 rate = 0.5,
                 balance_dir = 1,
                 balance_res = 0.5,
                 strength = 1,
                 reliable = 0.9,
                 export = TRUE,
                 patterns = FALSE,
                 slopes = FALSE)
gridSearch(data = data,
           yname = 'y_obs',
           tname = 't',
           pname = 'pid',
           pattern = 'contingency',
           steps = 1,
           reliability = 0.9,
           seed = 11234,
           workers = 10)
#> This function drops missing values. Tread carefully.
#>
#>  We now start the calculations. There will be 3 steps.
#>
#>  Step 1 for grid search...
#>
#>  Step 2 for grid search...
#>
#>  Step 3 for grid search...
#> # A tibble: 8,200 x 6
#>     rate strength direction       error pattern         n
#>    <dbl>    <dbl> <fct>           <dbl> <chr>       <int>
#>  1     0     0.05 100% Down        23.0 contingency    50
#>  2     0     0.05 75% Down-25% Up  23.8 contingency    50
#>  3     0     0.05 50% Down-50% Up  22.7 contingency    50
#>  4     0     0.05 25% Down-75% Up  24.0 contingency    50
#>  5     0     0.05 100% Up          24.4 contingency    50
#>  6     0     0.1  100% Down        23.1 contingency    50
#>  7     0     0.1  75% Down-25% Up  23.9 contingency    50
#>  8     0     0.1  50% Down-50% Up  22.8 contingency    50
#>  9     0     0.1  25% Down-75% Up  23.8 contingency    50
#> 10     0     0.1  100% Up          24.2 contingency    50
#> # i 8,190 more rows